simplify types, fix docs
